Add files via upload
This task was to use a for loop to print out each river's name, then to use another for loop to add up and total the length of all the rivers. Another task was to print out every river's name that begin with letter M and to print out every river's length in Kilometres!